CMake: win32: further fix for boost linker incompatibility
Package mingw-w64-x86_64-headers-git 10.0.0.r186.gfc55e181b-1
defines _WIN32_WINNT as 0x601, but boost is built against the r157
headers that defined this preprocessor as 0x603, leading to static
link issues.

This fixes the issue, but if a future version of boost is released,
it will be built against the r186 headers and therefore may require
further action to restore compatibility.