set concurrency limit for CI workflow
When a PR is synchronized it will trigger another workflow run, running in parallel to the previous run. This causes issues as we easily hit our GitHub limit of simultaneous jobs. Additionally, many jobs will fail if a force push occurs as they depend on the commit still existing at specific steps.