Skip irrelevant submodules when building on Arch
The ffmpeg submodules are quite large, so pulling them in when building on Arch adds some unnecessary extra time to the build. When building on Arch, we're only interested in `ffpmeg-linux-x86_64` and the others are not used at all.

Arguably we could conditionally pull `ffpmeg-linux-aarch64` instead of its x86_64 counterpart when building for / on aarch64, but because `arch` doesn't specify aarch64 as a valid architecture for this package, there's no real need.