feat: Add nightly build change log to index.html
This commit adds a change log section to the index.html file of the project. The change log displays the commit messages for the most recent commits on the "nightly" branch of the LizardByte/Sunshine repository. The maximum number of commits displayed is controlled by the "nightlyCommitLimit" property, which is set to 20. If the user is more than 20 commits behind the current version, a warning message is displayed. The commit messages are fetched using the GitHub API and are displayed in an unordered list.