Update input.cpp with the file path /Dev/uinput
The most common errors are

File not found. ??? What file? It doesn't say. So all I have to work with is some file presumably related to input emulation doesn't exist. It's better to say what file it is

Permission denied. Again no explanation of where

The error can be tackled directly if you tell me what the file is