build: Fix Docker cross-compiling for Ubuntu 22.04
Works for amd64, arm64, and ppc64le. Unlike Fedora, this does require
QEMU because APT seems incapable of operating on a different root
directory for a different architecture.

As with Fedora, cross-compiling CUDA works by borrowing the RHEL 8
packages because this is the only distro with NVIDIA packages for
ppc64le.