Link to Setup docs in README
It isn't clear how to install Sunshine on all operating systems just looking at the Downloads. macOS users, for instance, only see a Portfile, and won't know about the Homebrew package.

The documentation is mentioned at the top of the README, but it's easy to miss and there doesn't seem to be an easy way to get to the Sunshine-specific documentation from there.

A super-obvious "Installation" section that links to the Setup docs fixes this.