Add the ability to force an IDR from the troubleshooting menu.
Sometimes, devices are simply receiving video, and are not in locations which are amiable to retrieval.

For example, a device mounted on one's head, even if within an openable shell, is awkward to retrieve.

In these situations, higher resolutions and higher bitrates are often used, which leaves little room for any error correction or recovery; and the current mechanisms used for these can fail anyway.

In this case, it's required that there is an easy way to manually recover the stream.

While using the web UI is theoretically suboptimal for this, the exposure of an HTTP-based API makes it possible to automatically send IDR signals "every so often", or to provide some form of trigger for user IDR.

It is also possible that in some situations, a user may simply not wish to interrupt the flow of a stream; for example, when using Sunshine/Moonlight as part of a home theatre setup.

In this case, they can press the button on the web UI and playback can continue more-or-less uninterrupted.