Linux: simplify sleep before next frame snapshot
Before this commit kmsgrab, x11grab, cuda used a "busy" wait that did
not in fact keep the thread busy since it called `sleep_for()`.
The whole waiting period is now slept for at once.

Wlgrab used a true busy wait for a third of the waiting time. This
entailed a rather high CPU usage. The watiting method has been aligned
to the simplified method of the other Linux capture backends.

(Benchmark data can be found in the discussion of PR #2333.)