Bump MAX_FEC_BLOCKS to 4
Sunshine implementation arbitrarily limits itself to 3 FEC blocks even though the protocol supports 4. This is fine for most use cases. However, when dealing with large payloads (low FPS, high bitrate), there is a risk that the resulting packet size exceeds the capability of the error correction code.

This change removes magic constants to makes full use of the of the protocol.