fix(web-ui): fix build if ancestor directory of repository is a symlink
This commit addresses a situation where the user clones the Sunshine
directory in a location such that an ancestor of the repository is a
symlink. For example, imagine that the `~/git` directory is a symlink
to the `~/Documents/git` directory. Further imagine that the user clones
Sunshine into `~/git/Sunshine`. In this setup, attempting to build
the web-ui will fail with a series of incredibly cryptic error messages,
which look like this:

    RollupError: The "fileName" or "name" properties of emitted chunks
    and assets must be strings that are neither absolute nor relative
    paths, received "../../../../../../Documents/git/Sunshine/src_assets/common/assets/web/password.html".

This seems to occur because vite isn't aware that `Documents/git` and
`git` are actually the same directory as each other.

To fix the problem, this commit resolves the symlinks (if any) in
`vite.config.js` so that vite will only have to deal with one
canonical path, rather than two. This prevents the error when building.