Update Linux service to start after desktop target
Starting the Sunshine service at the same time as the desktop causes a race condition, this is currently solved with a delay.
I've added an After= dependency for the desktop, which should prevent the service from starting until the desktop has loaded. I've also commented out the delay, since this would not be needed.
This is working on my system (Fedora 41 with KDE), but I haven't tested other desktop environments.