pair from standard input working again
the text interface is broken for two years and nobody notices what else is new?I guess nobody use this it was broken. for two years apparently   I use a remote server with no
physical access and no web interface only text interface . when I used it two years ago it worked no big deal maybe a test would help but now I
can easier fix it the next time it breaks I know essentially how it
works took two days to find this learn how this works so I could fix it

I have to include an unrelated clang commit because otherwise I could
not test the build

i use android client all difference makes a difference. 99,99 percent is
a local rig with physical tele type and gleaming Trinitron and the whole
nine yards. shiny oak desk perhaps tasteful. this was not mentioned anywhere the logcat tag for the
android client . understand able since it is probably tested from a desk
not on the go~

~~~
logcat -d|ack LimeLog
~~~

 # remove the web interface from the build BUILD_WEBUI=OFF

this is an example of market pressure going too far. I understand the
enormous demand for computer operator assistance. when asked in 47 if
anyone could learn to use a computing machine von Neumann answered
absolutely. turns out no they Can not. so there is an enormous industry
for hand holding

if I could
exclude the web interface from the cmake build script it's just slightly provocative to have computers 101 force enabled   I would be the happiest man in the world

 # technical details

I hardly even remember but I think this is what happened. the client
stored the server state but not the server! so the next attempt the
client could never connect because the server was different. solved by
saving state even on failed pairing. and one time an error code was
returned together with paired=1 adding to the confusion. and something
more that I don't remember ready to sleep after 48 hours insomnia  and
migraine

oh right now I remember the return statement was missing probably
something changed for asynchronous pairing and a return statement was
never added leading to a successful pair and immediately a invalid
request message. strange that no one complained for two years I guess
they show each other how to do it and the web interface is vastly  preferred for struggling computer operators .
lonely text operator here it's lonely at the top. guess I have to post
this in Linux master race or something or text terminal master race

the original impetus was the never ending input problem on remote server
I tried to ask about it and absolutely zero people exactly a remote
machine with no physical access